requirements.txt
requests
google-cloud-storage
Pillow